requirements.txt
customtkinter
mcrcon
croniter
cryptography
python-dotenv
pystray
Pillow
pyinstaller
